feat: Observe and display movie genres with chips
Added setupGenresObserver in Fragment to observe genres from ViewModel and update UI with chips.
Implemented setupGenresChipGroups to populate the ChipGroup for both favorite and disliked genres.
Chips are created dynamically for each genre and added to corresponding ChipGroup.
Listeners are set for chips to ensure that a genre can only be selected as either "favorite" or "disliked" at a time.